feat: доавить поддержку лайков и удаления карточки

+++ src/utils/api.js
@@ -62,24 +62,26 @@ class Api {
             })
     }
 
-    putLike({ cardID }) {
-        return fetch(`${this._baseUrl}/cards/${cardID}/likes`, {
-            method: 'PUT',
-            headers: this._headers
-        })
-            .then((res) => {
-                return this._getResponseData(res)
+    changeLikeCardStatus({ cardID, isLiked }) {
+        if (isLiked) {
+            return fetch(`${this._baseUrl}/cards/${cardID}/likes`, {
+                method: 'DELETE',
+                headers: this._headers
             })
-    }
+                .then((res) => {
 
-    deleteLike({ cardID }) {
-        return fetch(`${this._baseUrl}/cards/${cardID}/likes`, {
-            method: 'DELETE',
-            headers: this._headers
-        })
-            .then((res) => {
-                return this._getResponseData(res)
+                    return this._getResponseData(res)
+                })
+        } else {
+            return fetch(`${this._baseUrl}/cards/${cardID}/likes`, {
+                method: 'PUT',
+                headers: this._headers
             })
+                .then((res) => {
+
+                    return this._getResponseData(res)
+                })
+        }
     }
 
     updateAvatar({ avatarLink }) {
@@ -96,13 +98,15 @@ class Api {
     }
 
     _getResponseData(res) {
-
         if (res.ok) {
+
             return res.json()
         }
         // если ошибка, отклоняем промис
         return Promise.reject(`Ошибка: ${res.status}`)
     }
+
+
 }
 
 const api = new Api({
